Stackcontent: All functions (except for StandardLearning) are executed correctly in offline mode and synchronize with the database when switching to online mode.

+++ lib/screens/edit_stack_screen.dart
@@ -1,6 +1,10 @@
+import 'dart:async';
+
+import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter/material.dart';
 import 'package:memo_dex_prototyp/screens/bottom_navigation_screen.dart';
 import 'package:memo_dex_prototyp/screens/stack_content_screen.dart';
+import '../services/local/file_handler.dart';
 import '../services/rest/rest_services.dart';
 import '../widgets/create_stack_form.dart';
 import '../widgets/delete_message_box.dart';
@@ -22,13 +26,56 @@ class EditStackScreen extends StatefulWidget {
 
 class _EditStackScreenState extends State<EditStackScreen> {
 
+  late StreamSubscription subscription;
+  FileHandler fileHandler = FileHandler();
+  bool online = true;
+
+  @override
+  void initState()
+  {
+    super.initState();
+    _checkInternetConnection();
+    subscription = Connectivity().onConnectivityChanged.listen((ConnectivityResult result)
+    async
+    {
+      _checkInternetConnection();
+    });
+  }
+
+  void _checkInternetConnection() async
+  {
+    ConnectivityResult connectivityResult = await Connectivity().checkConnectivity();
+    bool isConnected = (connectivityResult == ConnectivityResult.mobile || connectivityResult == ConnectivityResult.wifi);
+
+    if(isConnected == false)
+    {
+      online = false;
+    } else
+    {
+      online = true;
+    }
+  }
+
   void showDeleteMessageBox() {
     showDialog(
       context: context,
       builder: (BuildContext context) {
         return DeleteMessageBox(
           onDelete: () async {
-            await RestServices(context).deleteStack(1, widget.stackId);
+            if(online)
+            {
+              await RestServices(context).deleteStack(1, widget.stackId);
+            }else
+            {
+              if(widget.stackId is int)
+              {
+                await fileHandler.editItemById("allStacks", "stack_id", widget.stackId, {"is_deleted":1, "is_updated": 1});
+              }else
+              {
+                await fileHandler.editItemById("allLocalStacks", "stack_id", widget.stackId, {"is_deleted":1, "is_updated": 1});
+              }
+            }
+
             Navigator.pushReplacement(
               context,
               MaterialPageRoute(
@@ -40,6 +87,12 @@ class _EditStackScreenState extends State<EditStackScreen> {
         );
       },
     );
+  }
+
+  @override
+  void dispose() {
+    subscription.cancel();
+    super.dispose();
   }
 
   @override
